refactor(cli): extract daemon, status, and init command groups
Move the daemon-lifecycle (start/stop/restart, detached spawn, tracing
file rotation, ForegroundTeeGuard), status/health, and init/onboarding
command handlers out of main.rs into commands/{daemon,status,init}.rs.

main.rs now resolves all shared symbols through the command prelude glob
instead of explicit per-item imports, so it stays warning-clean as it
shrinks. Each new group is re-exported from the prelude so the dispatch
match and cross-group calls resolve.

Pure code move, no behavior change.
