+++ crates/librefang-cli/src/main.rs
@@ -31,20 +31,10 @@ mod tui;
 mod ui;
 
 use clap::Parser;
-use cli::*;
+// All other shared symbols (cli defs, common helpers, command groups, and the
+// std/external short names) come through the command prelude glob, which is
+// exempt from unused-import warnings as `main.rs` keeps shrinking.
 use commands::prelude::*;
-use colored::Colorize;
-use librefang_api::server::read_daemon_info;
-use librefang_extensions::dotenv;
-use librefang_kernel::{config::load_config, AgentSubsystemApi, LibreFangKernel, LlmSubsystemApi};
-use librefang_types::agent::{AgentId, AgentManifest};
-use std::ffi::OsString;
-use std::io::{self, BufRead, Write};
-#[cfg(unix)]
-use std::os::unix::io::RawFd;
-use std::path::PathBuf;
-use std::process::Stdio;
-use std::sync::atomic::AtomicBool;
 #[cfg(windows)]
 use std::sync::atomic::Ordering;
 use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
@@ -677,1233 +667,6 @@ fn main() {
             keep_config,
         }) => cmd_uninstall(confirm, keep_config),
         Some(Commands::HashPassword { password }) => cmd_hash_password(password),
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Commands
-// ---------------------------------------------------------------------------
-
-fn cmd_init(quick: bool) {
-    let home = match dirs::home_dir() {
-        Some(h) => h,
-        None => {
-            ui::error(&i18n::t("error-home-dir"));
-            std::process::exit(1);
-        }
-    };
-
-    let librefang_dir = cli_librefang_home();
-
-    // When an existing config is detected in interactive mode, redirect to the
-    // upgrade path so user settings (channels, keys, etc.) are preserved.
-    // The interactive wizard unconditionally overwrites config.toml, which
-    // would silently delete channels and custom configuration (#1862).
-    if !quick && librefang_dir.join("config.toml").exists() {
-        ui::hint("Existing installation detected — running upgrade to preserve your settings.");
-        ui::hint("To start fresh, remove ~/.librefang/config.toml and run `librefang init` again.");
-        cmd_init_upgrade();
-        return;
-    }
-
-    // --- Ensure directories exist ---
-    if !librefang_dir.exists() {
-        std::fs::create_dir_all(&librefang_dir).unwrap_or_else(|e| {
-            ui::error_with_fix(
-                &i18n::t_args(
-                    "error-create-dir",
-                    &[("path", &librefang_dir.display().to_string())],
-                ),
-                &i18n::t_args(
-                    "error-create-dir-fix",
-                    &[("path", &home.display().to_string())],
-                ),
-            );
-            eprintln!("  {e}");
-            std::process::exit(1);
-        });
-        restrict_dir_permissions(&librefang_dir);
-    }
-
-    let data_dir = librefang_dir.join("data");
-    if !data_dir.exists() {
-        std::fs::create_dir_all(&data_dir).unwrap_or_else(|e| {
-            eprintln!("Error creating data dir: {e}");
-            std::process::exit(1);
-        });
-    }
-
-    // Sync registry content (downloads to registry/, pre-installs providers/integrations/assistant)
-    librefang_runtime::registry_sync::sync_registry(
-        &librefang_dir,
-        librefang_runtime::registry_sync::DEFAULT_CACHE_TTL_SECS,
-        "",
-    );
-
-    // Initialize vault if not already initialized
-    init_vault_if_missing(&librefang_dir);
-
-    // Initialize git repo for config version control
-    init_git_if_missing(&librefang_dir);
-
-    if quick {
-        cmd_init_quick(&librefang_dir);
-    } else if !std::io::IsTerminal::is_terminal(&std::io::stdin())
-        || !std::io::IsTerminal::is_terminal(&std::io::stdout())
-    {
-        ui::hint(&i18n::t("hint-non-interactive"));
-        ui::hint(&i18n::t("hint-non-interactive-wizard"));
-        cmd_init_quick(&librefang_dir);
-    } else {
-        cmd_init_interactive(&librefang_dir);
-    }
-
-    // Fallback: ensure config.toml exists even if wizard was cancelled/failed
-    let config_path = librefang_dir.join("config.toml");
-    if !config_path.exists() {
-        let (provider, api_key_env, model) = detect_best_provider();
-        write_config_if_missing(&librefang_dir, &provider, &model, &api_key_env);
-    }
-}
-
-/// Upgrade an existing LibreFang installation: backup config, sync registry, merge new defaults.
-fn cmd_init_upgrade() {
-    let librefang_dir = cli_librefang_home();
-    let config_path = librefang_dir.join("config.toml");
-
-    // 1. Must have an existing installation
-    if !config_path.exists() {
-        ui::error("Nothing to upgrade — no config.toml found. Run `librefang init` first.");
-        std::process::exit(1);
-    }
-
-    ui::banner();
-    ui::blank();
-    ui::section("Upgrading LibreFang installation");
-
-    // Four upgrade steps: backup, registry sync, vault/git, config merge.
-    let mut p = progress::auto("Upgrading", Some(4));
-
-    // 2. Backup existing config under backups/ (keep last 3)
-    p.set_message("Backing up config");
-    let backups_dir = librefang_dir.join("backups");
-    if let Err(e) = std::fs::create_dir_all(&backups_dir) {
-        p.finish_with_failure(&format!("Failed to create backups dir: {e}"));
-        std::process::exit(1);
-    }
-    let backup_name = format!("config-{}.toml", format_local_timestamp());
-    let backup_path = backups_dir.join(&backup_name);
-    if let Err(e) = std::fs::copy(&config_path, &backup_path) {
-        p.finish_with_failure(&format!("Failed to backup config: {e}"));
-        std::process::exit(1);
-    }
-    restrict_file_permissions(&backup_path);
-    prune_old_config_backups(&backups_dir, 3);
-    p.tick(1);
-    ui::success(&format!("Backed up config to backups/{backup_name}"));
-
-    // 3. Sync registry (TTL=0 forces refresh regardless of last sync time)
-    p.set_message("Syncing registry");
-    if librefang_runtime::registry_sync::sync_registry(&librefang_dir, 0, "") {
-        p.tick(1);
-        ui::success("Registry synced");
-    } else {
-        p.tick(1);
-        ui::hint("Registry sync failed (network issue?) — continuing with cached content");
-    }
-
-    // 4. Ensure data dir, vault, and git exist
-    p.set_message("Initialising vault/git");
-    let data_dir = librefang_dir.join("data");
-    if !data_dir.exists() {
-        let _ = std::fs::create_dir_all(&data_dir);
-    }
-    init_vault_if_missing(&librefang_dir);
-    init_git_if_missing(&librefang_dir);
-
-    // Ensure .gitignore excludes the backups/ directory (may be missing in older installations)
-    let gitignore = librefang_dir.join(".gitignore");
-    if gitignore.exists() {
-        if let Ok(content) = std::fs::read_to_string(&gitignore) {
-            if !content.lines().any(|l| l.trim() == "backups/") {
-                let _ = std::fs::write(&gitignore, format!("{content}backups/\n"));
-            }
-        }
-    }
-    p.tick(1);
-
-    // 5. Merge new default config fields
-    p.set_message("Merging config fields");
-    let existing_raw = match std::fs::read_to_string(&config_path) {
-        Ok(s) => s,
-        Err(e) => {
-            p.finish_with_failure(&format!("Upgrade aborted: failed to read config.toml: {e}"));
-            std::process::exit(1);
-        }
-    };
-
-    let existing: toml::Value = match toml::from_str(&existing_raw) {
-        Ok(v) => v,
-        Err(e) => {
-            p.finish_with_failure(&format!(
-                "Upgrade aborted: failed to parse config.toml: {e}"
-            ));
-            ui::hint(&format!(
-                "Your original config was saved to backups/{backup_name}"
-            ));
-            std::process::exit(1);
-        }
-    };
-
-    let (provider, api_key_env, model) = detect_best_provider();
-    let default_config_str = render_init_default_config(&provider, &model, &api_key_env);
-    let defaults: toml::Value = match toml::from_str(&default_config_str) {
-        Ok(v) => v,
-        Err(e) => {
-            p.finish_with_failure(&format!(
-                "Upgrade aborted: failed to parse default config template: {e}"
-            ));
-            std::process::exit(1);
-        }
-    };
-
-    // Find top-level keys/sections missing from user config and append them
-    // as TOML fragments. This preserves the original file's comments and formatting.
-    let added = find_missing_toplevel_keys(&existing, &defaults);
-
-    if added.is_empty() {
-        ui::success("Config is already up to date — no new fields added");
-    } else {
-        // Partition into scalars (must stay in TOML root scope) and tables.
-        // Scalars appended after a [table] header would be absorbed into that
-        // table's scope, potentially colliding with same-named sub-keys (#2021).
-        let (scalar_keys, table_keys): (Vec<_>, Vec<_>) = added
-            .iter()
-            .partition(|k| defaults.get(*k).is_none_or(|v| !v.is_table()));
-
-        let mut content = existing_raw.clone();
-
-        // Insert scalar keys before the first [table] header so they remain
-        // top-level in the TOML document.
-        if !scalar_keys.is_empty() {
-            let mut scalar_snippet = String::new();
-            for key in &scalar_keys {
-                if let Some(val) = defaults.get(*key) {
-                    let mut fragment = toml::map::Map::new();
-                    fragment.insert((*key).clone(), val.clone());
-                    if let Ok(s) = toml::to_string_pretty(&toml::Value::Table(fragment)) {
-                        scalar_snippet.push_str(&s);
-                    }
-                }
-            }
-            // Find the first line that starts with '[' (a table header).
-            // We search for "\n[" then insert just before the '['.
-            if let Some(pos) = content.find("\n[").map(|p| p + 1) {
-                content.insert_str(pos, &format!("{scalar_snippet}\n"));
-            } else {
-                // No table headers in file — appending is safe.
-                content.push('\n');
-                content.push_str(&scalar_snippet);
-            }
-        }
-
-        // Append table sections at the end of the file.
-        if !table_keys.is_empty() {
-            content.push_str("\n# ── Added by upgrade ────────────────────────────────────\n");
-            for key in &table_keys {
-                if let Some(val) = defaults.get(*key) {
-                    let mut fragment = toml::map::Map::new();
-                    fragment.insert((*key).clone(), val.clone());
-                    if let Ok(snippet) = toml::to_string_pretty(&toml::Value::Table(fragment)) {
-                        content.push('\n');
-                        content.push_str(&snippet);
-                    }
-                }
-            }
-        }
-
-        if let Err(e) = std::fs::write(&config_path, &content) {
-            p.finish_with_failure(&format!("Upgrade aborted: failed to write config: {e}"));
-            ui::hint(&format!(
-                "Your original config was saved to backups/{backup_name}"
-            ));
-            std::process::exit(1);
-        }
-        restrict_file_permissions(&config_path);
-        ui::success(&format!("Added {} new config section(s):", added.len()));
-        for key in &added {
-            ui::kv("  +", key);
-        }
-    }
-    p.tick(1);
-    p.finish("Upgrade steps complete");
-
-    // 6. Check for legacy ~/.openclaw installation
-    if let Some(home) = dirs::home_dir() {
-        let openclaw_dir = home.join(".openclaw");
-        if openclaw_dir.exists() {
-            ui::blank();
-            ui::hint("Legacy ~/.openclaw installation detected.");
-            ui::hint("Run `librefang migrate --from openclaw` to migrate your data.");
-        }
-    }
-
-    // 7. Warn users whose require_approval list predates the file_write default (#1861).
-    // The default was expanded to include file_write and file_delete, but users who
-    // had an explicit `require_approval = [...]` entry in their config won't pick up
-    // the new default automatically.
-    let approval_needs_update = existing
-        .get("approval")
-        .and_then(|a| a.get("require_approval"))
-        .and_then(|r| r.as_array())
-        .is_some_and(|list| {
-            let has_shell = list.iter().any(|v| v.as_str() == Some("shell_exec"));
-            let missing_new = ["file_write", "file_delete", "apply_patch"]
-                .iter()
-                .any(|tool| !list.iter().any(|v| v.as_str() == Some(*tool)));
-            has_shell && missing_new
-        });
-    if approval_needs_update {
-        ui::blank();
-        ui::hint(
-            "Your require_approval list only contains \"shell_exec\". \
-             File operations (file_write, file_delete) now require approval by default.",
-        );
-        ui::hint(
-            "To enable: add \"file_write\" and \"file_delete\" to require_approval in config.toml",
-        );
-    }
-
-    // 8. Summary
-    ui::blank();
-    ui::success("Upgrade complete!");
-    ui::kv("Backup", &format!("backups/{backup_name}"));
-    if !added.is_empty() {
-        ui::kv("New fields", &added.len().to_string());
-    }
-    ui::blank();
-}
-
-/// Keep only the `keep` most recent `config-*.toml` backups under `backups_dir`.
-/// The embedded `YYYYMMDD-HHMMSS` timestamp sorts lexicographically, so a
-/// filename sort gives the same order as a chronological sort.
-fn prune_old_config_backups(backups_dir: &std::path::Path, keep: usize) {
-    let Ok(entries) = std::fs::read_dir(backups_dir) else {
-        return;
-    };
-    let mut files: Vec<std::path::PathBuf> = entries
-        .flatten()
-        .filter_map(|e| {
-            let path = e.path();
-            let name = path.file_name()?.to_str()?;
-            if name.starts_with("config-") && name.ends_with(".toml") {
-                Some(path)
-            } else {
-                None
-            }
-        })
-        .collect();
-    files.sort();
-    if files.len() > keep {
-        for old in &files[..files.len() - keep] {
-            let _ = std::fs::remove_file(old);
-        }
-    }
-}
-
-/// Find top-level keys in `defaults` that are missing from `existing`.
-/// Only checks top-level — does not recurse into sub-tables to avoid
-/// injecting partial sections the user intentionally omitted.
-fn find_missing_toplevel_keys(existing: &toml::Value, defaults: &toml::Value) -> Vec<String> {
-    let (Some(existing_table), Some(defaults_table)) = (existing.as_table(), defaults.as_table())
-    else {
-        return Vec::new();
-    };
-    defaults_table
-        .keys()
-        .filter(|k| !existing_table.contains_key(*k))
-        .cloned()
-        .collect()
-}
-
-/// Initialize vault if it doesn't exist yet (silent no-op if already initialized).
-fn init_vault_if_missing(librefang_dir: &std::path::Path) {
-    let vault_path = librefang_dir.join("vault.enc");
-    if vault_path.exists() {
-        return; // Already initialized
-    }
-
-    let mut vault = librefang_extensions::vault::CredentialVault::new(vault_path);
-    if let Err(e) = vault.init() {
-        // Silently skip vault init on failure - it's optional
-        tracing::debug!("vault init skipped: {e}");
-    }
-}
-
-/// Initialize a git repo in ~/.librefang/ for config version control.
-fn init_git_if_missing(librefang_dir: &std::path::Path) {
-    if librefang_dir.join(".git").exists() {
-        return;
-    }
-
-    let Ok(status) = std::process::Command::new("git")
-        .args(["init", "-q", "-b", "main"])
-        .current_dir(librefang_dir)
-        .status()
-    else {
-        tracing::debug!("git not available, skipping repo init");
-        return;
-    };
-    if !status.success() {
-        tracing::debug!("git init failed");
-        return;
-    }
-
-    // Write .gitignore for sensitive/temporary files
-    let gitignore = librefang_dir.join(".gitignore");
-    if !gitignore.exists() {
-        let _ = std::fs::write(
-            &gitignore,
-            "secrets.env\nvault.enc\ndaemon.json\nlogs/\ncache/\nregistry/\ndata/\nbackups/\n*.db\n*.db-shm\n*.db-wal\n",
-        );
-    }
-
-    // Initial commit
-    let _ = std::process::Command::new("git")
-        .args(["add", "-A"])
-        .current_dir(librefang_dir)
-        .status();
-    let _ = std::process::Command::new("git")
-        .args(["commit", "-q", "-m", "chore: initial librefang config"])
-        .current_dir(librefang_dir)
-        .status();
-}
-
-/// Quick init: no prompts, auto-detect, write config + .env, print next steps.
-fn cmd_init_quick(librefang_dir: &std::path::Path) {
-    ui::banner();
-    ui::blank();
-
-    let (provider, api_key_env, model) = detect_best_provider();
-
-    write_config_if_missing(librefang_dir, &provider, &model, &api_key_env);
-
-    ui::blank();
-    ui::success(&i18n::t("init-quick-success"));
-    ui::kv(&i18n::t("label-provider"), &provider);
-    ui::kv(&i18n::t("label-model"), &model);
-    ui::blank();
-    ui::next_steps(&[&i18n::t("init-next-start"), &i18n::t("init-next-chat")]);
-}
-
-/// Interactive 5-step onboarding wizard (ratatui TUI).
-fn cmd_init_interactive(librefang_dir: &std::path::Path) {
-    use tui::screens::init_wizard::{self, InitResult, LaunchChoice};
-
-    match init_wizard::run() {
-        InitResult::Completed {
-            provider,
-            model,
-            daemon_started,
-            launch,
-        } => {
-            // Print summary after TUI restores terminal
-            ui::blank();
-            ui::success(&i18n::t("init-interactive-success"));
-            ui::kv(&i18n::t("label-provider"), &provider);
-            ui::kv(&i18n::t("label-model"), &model);
-
-            if daemon_started {
-                ui::kv_ok(&i18n::t("label-daemon"), "running");
-            }
-            ui::blank();
-
-            // Execute the user's chosen launch action.
-            match launch {
-                LaunchChoice::Desktop => {
-                    launch_desktop_app(librefang_dir);
-                }
-                LaunchChoice::Dashboard => {
-                    if let Some(base) = find_daemon() {
-                        let url = format!("{base}/");
-                        ui::success(&i18n::t_args("dashboard-opening", &[("url", &url)]));
-                        if !open_in_browser(&url) {
-                            ui::hint(&i18n::t_args(
-                                "hint-could-not-open-browser-visit",
-                                &[("url", &url)],
-                            ));
-                        }
-                    } else {
-                        ui::error(&i18n::t("daemon-not-running-start"));
-                    }
-                }
-                LaunchChoice::Chat => {
-                    ui::hint(&i18n::t("hint-starting-chat"));
-                    ui::blank();
-                    // Note: tracing was initialized for stderr (init is a CLI
-                    // subcommand).  The chat TUI takes over the terminal with
-                    // raw mode so stderr output is suppressed.  We can't
-                    // reinitialize tracing (global subscriber is set once).
-                    cmd_quick_chat(None, None);
-                }
-            }
-        }
-        InitResult::Cancelled => {
-            println!("  {}", i18n::t("init-cancelled"));
-        }
-    }
-}
-
-/// Launch the librefang-desktop Tauri app, connecting to the running daemon.
-fn launch_desktop_app(_librefang_dir: &std::path::Path) {
-    if let Some(path) = desktop_install::find_desktop_binary() {
-        desktop_install::launch(&path);
-        return;
-    }
-
-    // Not installed — offer to download
-    if let Some(installed) = desktop_install::prompt_and_install() {
-        desktop_install::launch(&installed);
-    }
-}
-
-/// Auto-detect the best available provider.
-///
-/// Delegates to the runtime's `detect_available_provider()` which probes 13+
-/// providers (OpenAI, Anthropic, Gemini, Groq, DeepSeek, OpenRouter, Mistral,
-/// Together, Fireworks, xAI, Perplexity, Cohere, Azure OpenAI) plus the
-/// GOOGLE_API_KEY alias.  Falls back to local Ollama, then the interactive
-/// free-provider TUI guide.
-fn detect_best_provider() -> (String, String, String) {
-    // 1. Check all cloud provider API keys via the runtime registry
-    if let Some((provider, _model, env_var)) =
-        librefang_runtime::drivers::detect_available_provider()
-    {
-        // Capitalize provider name for display (e.g. "groq" → "Groq")
-        let display_name = {
-            let mut c = provider.chars();
-            match c.next() {
-                None => String::new(),
-                Some(first) => first.to_uppercase().to_string() + c.as_str(),
-            }
-        };
-        // CLI-backed providers return an empty env_var (auth via OAuth token
-        // or keychain, not an env variable). Display a readable placeholder
-        // so the i18n message doesn't end with an empty parenthetical.
-        let auth_display = if env_var.is_empty() {
-            "CLI login"
-        } else {
-            env_var
-        };
-        ui::success(&i18n::t_args(
-            "detected-provider",
-            &[("display", &display_name), ("env_var", auth_display)],
-        ));
-        return (
-            provider.to_string(),
-            env_var.to_string(),
-            default_model_for_provider(provider),
-        );
-    }
-
-    // 2. Check if Ollama is running locally (no API key needed)
-    if check_ollama_available() {
-        ui::success(&i18n::t("detected-ollama"));
-        return (
-            "ollama".to_string(),
-            "OLLAMA_API_KEY".to_string(),
-            default_model_for_provider("ollama"),
-        );
-    }
-
-    // 3. No API key found — launch TUI guide to pick a free provider
-    {
-        if let Some(result) = guide_free_provider_setup() {
-            return result;
-        }
-    }
-
-    // 4. Non-interactive fallback: just print hints
-    ui::hint(&i18n::t("hint-no-api-keys"));
-    ui::hint(&i18n::t("hint-groq-free"));
-    ui::hint(&i18n::t("hint-gemini-free"));
-    ui::hint(&i18n::t("hint-deepseek-free"));
-    ui::hint(&i18n::t("hint-ollama-local"));
-    (
-        "groq".to_string(),
-        "GROQ_API_KEY".to_string(),
-        default_model_for_provider("groq"),
-    )
-}
-
-/// Interactive TUI guide: help user pick a free LLM provider and set up an API key.
-/// Returns `Some((provider, env_var, model))` on success, `None` if user cancels.
-fn guide_free_provider_setup() -> Option<(String, String, String)> {
-    use tui::screens::free_provider_guide::{self, GuideResult};
-
-    match free_provider_guide::run() {
-        GuideResult::Completed { provider, env_var } => {
-            ui::success(&i18n::t_args("config-saved-key", &[("env_var", &env_var)]));
-            let model = default_model_for_provider(&provider);
-            Some((provider, env_var, model))
-        }
-        GuideResult::Skipped => None,
-    }
-}
-
-/// Quick probe to check if Ollama is running on localhost.
-fn check_ollama_available() -> bool {
-    std::net::TcpStream::connect_timeout(
-        &std::net::SocketAddr::from(([127, 0, 0, 1], 11434)),
-        std::time::Duration::from_millis(500),
-    )
-    .is_ok()
-}
-
-fn render_init_default_config(provider: &str, model: &str, api_key_env: &str) -> String {
-    INIT_DEFAULT_CONFIG_TEMPLATE
-        .replace("{{provider}}", provider)
-        .replace("{{model}}", model)
-        .replace("{{api_key_env}}", api_key_env)
-}
-
-fn default_model_for_provider(provider: &str) -> String {
-    let catalog = librefang_runtime::model_catalog::ModelCatalog::default();
-    catalog
-        .default_model_for_provider(provider)
-        .unwrap_or_else(|| "local-model".to_string())
-}
-
-/// Write config.toml if it doesn't already exist.
-fn write_config_if_missing(
-    librefang_dir: &std::path::Path,
-    provider: &str,
-    model: &str,
-    api_key_env: &str,
-) {
-    let config_path = librefang_dir.join("config.toml");
-    if config_path.exists() {
-        ui::check_ok(&i18n::t_args(
-            "error-config-exists",
-            &[("path", &config_path.display().to_string())],
-        ));
-    } else {
-        let default_config = render_init_default_config(provider, model, api_key_env);
-        std::fs::write(&config_path, &default_config).unwrap_or_else(|e| {
-            ui::error_with_fix(&i18n::t("error-write-config"), &e.to_string());
-            std::process::exit(1);
-        });
-        restrict_file_permissions(&config_path);
-        ui::success(&i18n::t_args(
-            "error-config-created",
-            &[("path", &config_path.display().to_string())],
-        ));
-    }
-
-    // Write config.example.toml with the full annotated template for reference
-    let example_path = librefang_dir.join("config.example.toml");
-    if !example_path.exists() {
-        let example_content = include_str!("../templates/init_default_config.toml");
-        if let Err(e) = std::fs::write(&example_path, example_content) {
-            ui::hint(&format!("Could not write config.example.toml: {e}"));
-        }
-    }
-}
-
-fn daemon_log_path_for_home(home_dir: &std::path::Path) -> PathBuf {
-    home_dir.join("logs").join("daemon.log")
-}
-
-fn daemon_log_path_for_config(config: Option<&std::path::Path>) -> PathBuf {
-    let daemon = daemon_config_context(config);
-    if let Some(ref log_dir) = daemon.log_dir {
-        log_dir.join("daemon.log")
-    } else {
-        daemon_log_path_for_home(&daemon.home_dir)
-    }
-}
-
-fn detached_daemon_args(config: Option<&std::path::Path>) -> Vec<OsString> {
-    let mut args = Vec::new();
-    if let Some(path) = config {
-        args.push(OsString::from("--config"));
-        args.push(path.as_os_str().to_owned());
-    }
-    args.push(OsString::from("start"));
-    args.push(OsString::from("--spawned"));
-    args
-}
-
-fn spawn_detached_daemon(
-    config: Option<&std::path::Path>,
-    log_path: &std::path::Path,
-) -> Result<std::process::Child, String> {
-    let exe = std::env::current_exe().map_err(|e| format!("resolve current executable: {e}"))?;
-    if let Some(log_dir) = log_path.parent() {
-        std::fs::create_dir_all(log_dir)
-            .map_err(|e| format!("create log directory {}: {e}", log_dir.display()))?;
-        restrict_dir_permissions(log_dir);
-    }
-
-    let stdout = std::fs::OpenOptions::new()
-        .create(true)
-        .append(true)
-        .open(log_path)
-        .map_err(|e| format!("open daemon log {}: {e}", log_path.display()))?;
-    let stderr = stdout
-        .try_clone()
-        .map_err(|e| format!("clone daemon log handle {}: {e}", log_path.display()))?;
-
-    let mut command = std::process::Command::new(exe);
-    command
-        .args(detached_daemon_args(config))
-        .stdin(Stdio::null())
-        .stdout(Stdio::from(stdout))
-        .stderr(Stdio::from(stderr))
-        .envs(std::env::vars());
-
-    #[cfg(unix)]
-    {
-        use std::os::unix::process::CommandExt;
-
-        unsafe {
-            command.pre_exec(|| {
-                if libc::setsid() == -1 {
-                    return Err(std::io::Error::last_os_error());
-                }
-                Ok(())
-            });
-        }
-    }
-
-    #[cfg(windows)]
-    {
-        use std::os::windows::process::CommandExt;
-
-        const DETACHED_PROCESS: u32 = 0x0000_0008;
-        const CREATE_NEW_PROCESS_GROUP: u32 = 0x0000_0200;
-        const CREATE_NO_WINDOW: u32 = 0x0800_0000;
-
-        command.creation_flags(DETACHED_PROCESS | CREATE_NEW_PROCESS_GROUP | CREATE_NO_WINDOW);
-    }
-
-    command
-        .spawn()
-        .map_err(|e| format!("spawn detached daemon: {e}"))
-}
-
-/// Generate a daily log path for the current daemon start.
-/// Returns e.g. ~/.librefang/logs/daemon-2026-04-23.log
-/// Same day restarts reuse the same file.
-fn timestamped_log_path(config: Option<&std::path::Path>) -> std::path::PathBuf {
-    let daemon = daemon_config_context(config);
-    let log_dir = daemon
-        .log_dir
-        .unwrap_or_else(|| daemon.home_dir.join("logs"));
-    let date = chrono_lite_date();
-    log_dir.join(format!("daemon-{date}.log"))
-}
-
-/// Prune rotated daemon logs older than `max_age_days`, keeping the log dir tidy.
-fn prune_rotated_logs(config: Option<&std::path::Path>, max_age_days: u64) {
-    let daemon = daemon_config_context(config);
-    let log_dir = daemon
-        .log_dir
-        .unwrap_or_else(|| daemon.home_dir.join("logs"));
-    let cutoff = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap()
-        .as_secs()
-        .saturating_sub(max_age_days.saturating_mul(86400));
-
-    let entries = match std::fs::read_dir(&log_dir) {
-        Ok(e) => e,
-        Err(_) => return,
-    };
-    for entry in entries.flatten() {
-        let path = entry.path();
-        let name = match path.file_name().and_then(|n| n.to_str()) {
-            Some(n) => n,
-            None => continue,
-        };
-        if !name.starts_with("daemon-") || !name.ends_with(".log") {
-            continue;
-        }
-        // Parse date from filename: daemon-YYYY-MM-DD.log
-        let date_str = name
-            .strip_prefix("daemon-")
-            .and_then(|s| s.strip_suffix(".log"));
-        let is_old = date_str
-            .and_then(parse_daily_date_timestamp)
-            .map(|ts| ts < cutoff)
-            .unwrap_or(false);
-        if is_old {
-            let _ = std::fs::remove_file(&path);
-        }
-    }
-}
-
-/// Guard that tees all stdout/stderr to a log file in foreground mode.
-/// On drop, restores original stdout/stderr and joins the tee thread.
-#[cfg(unix)]
-struct ForegroundTeeGuard {
-    _pipe_fd: RawFd, // kept alive to keep pipe open until guard drops
-}
-
-#[cfg(unix)]
-impl Drop for ForegroundTeeGuard {
-    fn drop(&mut self) {
-        // Restore original stdout/stderr
-        unsafe {
-            libc::dup2(self._pipe_fd, libc::STDOUT_FILENO);
-            libc::dup2(self._pipe_fd, libc::STDERR_FILENO);
-            libc::close(self._pipe_fd);
-        }
-    }
-}
-
-/// Set up tee for --foreground mode: redirect stdout/stderr to a pipe,
-/// spawn a background thread that copies to both terminal and log file.
-#[cfg(unix)]
-fn setup_foreground_tee(log_path: &std::path::Path) -> ForegroundTeeGuard {
-    // Ensure the parent directory exists (e.g. `~/.librefang/logs/`) before we
-    // try to open the log file. Fresh installations and test environments may
-    // not have this directory yet.
-    if let Some(parent) = log_path.parent() {
-        if let Err(e) = std::fs::create_dir_all(parent) {
-            // Report the failure on the real stderr (not yet redirected), then
-            // exit instead of limping forward into the silent-hang regime the
-            // old ordering produced.
-            eprintln!("Failed to create log directory {}: {e}", parent.display());
-            std::process::exit(1);
-        }
-    }
-
-    // Open the log file BEFORE redirecting stdout/stderr. If the open panics
-    // (permissions, read-only fs, parent still missing, …) the panic message
-    // reaches the real stderr the user is watching. Previously we opened the
-    // file AFTER `dup2`, so a failure wrote its panic message into a pipe
-    // whose reader hadn't been spawned yet — the message was trapped in the
-    // pipe buffer and the process appeared to hang at "Starting daemon…".
-    let log_file = std::sync::Mutex::new(
-        std::fs::OpenOptions::new()
-            .create(true)
-            .append(true)
-            .open(log_path)
-            .unwrap_or_else(|e| {
-                eprintln!("Failed to open daemon log file {}: {e}", log_path.display());
-                std::process::exit(1);
-            }),
-    );
-
-    // Create pipe for stdout+stderr (we'll write to it, background thread reads)
-    let mut fds = [0i32, 0i32];
-    unsafe {
-        libc::pipe(fds.as_mut_ptr());
-    }
-    let pipe_write = fds[1];
-    let pipe_read = fds[0];
-
-    // Save a copy of original stdout (to restore both fd 1 and fd 2 on
-    // drop). We don't keep a separate stderr copy: by the time the tee
-    // thread reads from the pipe, stdout and stderr have already been
-    // merged at the fd level, so we cannot route output back to the
-    // correct original fd. Writing to both copies would simply duplicate
-    // every line in any consumer that captures both fds (e.g. the Docker
-    // log driver), which is the bug this code path used to cause.
-    let stdout_copy = unsafe { libc::dup(libc::STDOUT_FILENO) };
-
-    // Redirect stdout and stderr to the pipe. From here on any write to the
-    // standard streams goes through the pipe and must be drained by the
-    // read thread below — do not fail between this point and `thread::spawn`.
-    unsafe {
-        libc::dup2(pipe_write, libc::STDOUT_FILENO);
-        libc::dup2(pipe_write, libc::STDERR_FILENO);
-        libc::close(pipe_write);
-    }
-
-    // Spawn background thread that reads from pipe and writes to both terminal and log
-    std::thread::spawn(move || {
-        let mut buf = [0u8; 8192];
-        loop {
-            let n =
-                unsafe { libc::read(pipe_read, buf.as_mut_ptr() as *mut libc::c_void, buf.len()) };
-            if n <= 0 {
-                // EOF or error — exit; guard Drop will restore stdout/stderr
-                unsafe { libc::close(pipe_read) };
-                break;
-            }
-            // Write to the saved stdout fd once. See comment at the dup site
-            // for why we don't also write to a stderr copy.
-            unsafe {
-                libc::write(stdout_copy, buf.as_ptr() as *const libc::c_void, n as usize);
-            }
-            // Write to log file
-            if let Ok(mut f) = log_file.lock() {
-                let _ = f.write_all(&buf[..n as usize]);
-                let _ = f.flush();
-            }
-        }
-        // guard Drop closes stdout_copy; pipe_read is closed above on break
-    });
-
-    ForegroundTeeGuard {
-        _pipe_fd: stdout_copy,
-    }
-}
-
-/// Ensure LibreFang is initialized (config.toml exists). Auto-runs quick init on first run.
-fn ensure_initialized(config: &Option<PathBuf>) {
-    match config {
-        None => {
-            let home = cli_librefang_home();
-            if !home.join("config.toml").exists() {
-                ui::hint("First run detected — running quick setup...");
-                cmd_init(true);
-            }
-        }
-        Some(path) => {
-            if !path.exists() {
-                ui::error_with_fix(
-                    &format!("Config file not found: {}", path.display()),
-                    "Run `librefang init` to create a default config at ~/.librefang/config.toml, or check the --config path.",
-                );
-                std::process::exit(1);
-            }
-        }
-    }
-}
-
-fn cmd_start(config: Option<PathBuf>, tail: bool, spawned: bool, foreground: bool) {
-    ensure_initialized(&config);
-
-    // Issue #5186 follow-up: `cmd_start` boots a real daemon, so a bad
-    // `config.toml` must abort here BEFORE `daemon_config_context` swallows
-    // the load error and substitutes `KernelConfig::default()`. The
-    // tolerant default would silently change `home_dir` (used a few lines
-    // below to detect an already-running daemon) and the spawned child
-    // would only fail-closed during its own boot — losing the field-name
-    // diagnostic from the parent's stderr in the process.
-    //
-    // `load_config` already prints the underlying error to stderr (so
-    // operators see the field name even before the tracing subscriber is
-    // wired up); we only need to short-circuit with a non-zero exit.
-    if load_config(config.as_deref()).is_err() {
-        std::process::exit(1);
-    }
-
-    let daemon = daemon_config_context(config.as_deref());
-    if let Some(base) = find_daemon_in_home(&daemon.home_dir) {
-        ui::error_with_fix(
-            &i18n::t_args("daemon-already-running", &[("url", &base)]),
-            &i18n::t("daemon-already-running-fix"),
-        );
-        std::process::exit(1);
-    }
-
-    if !spawned && !foreground {
-        let log_path = daemon_log_path_for_config(config.as_deref());
-        let mut child = spawn_detached_daemon(config.as_deref(), &log_path).unwrap_or_else(|e| {
-            ui::error_with_fix(&i18n::t("daemon-launch-fail"), &e);
-            std::process::exit(1);
-        });
-
-        let deadline = Instant::now() + Duration::from_secs(10);
-        loop {
-            if let Some(base) = find_daemon_in_home(&daemon.home_dir) {
-                let pid = child.id();
-                std::mem::forget(child);
-                ui::success(&i18n::t("daemon-started-bg"));
-                ui::kv(&i18n::t("label-pid"), &pid.to_string());
-                ui::kv(&i18n::t("label-api"), &base);
-                ui::kv(&i18n::t("label-dashboard"), &format!("{base}/"));
-                ui::kv(&i18n::t("label-log"), &log_path.display().to_string());
-                if tail {
-                    ui::hint(&i18n::t("hint-tail-stop"));
-                    ui::blank();
-                    show_log_file(&log_path, 50, true);
-                } else {
-                    ui::hint(&i18n::t("hint-stop-daemon"));
-                }
-                return;
-            }
-
-            match child.try_wait() {
-                Ok(Some(status)) => {
-                    ui::error_with_fix(
-                        &i18n::t_args("daemon-bg-exited", &[("status", &status.to_string())]),
-                        &i18n::t_args(
-                            "daemon-bg-exited-fix",
-                            &[("path", &log_path.display().to_string())],
-                        ),
-                    );
-                    std::process::exit(1);
-                }
-                Ok(None) => {}
-                Err(e) => {
-                    ui::error_with_fix(
-                        &i18n::t("daemon-bg-wait-fail"),
-                        &i18n::t_args(
-                            "daemon-bg-wait-fail-fix",
-                            &[
-                                ("error", &e.to_string()),
-                                ("path", &log_path.display().to_string()),
-                            ],
-                        ),
-                    );
-                    std::process::exit(1);
-                }
-            }
-
-            if Instant::now() >= deadline {
-                let pid = child.id();
-                std::mem::forget(child);
-                ui::success(&i18n::t("daemon-still-starting"));
-                ui::kv(&i18n::t("label-pid"), &pid.to_string());
-                ui::kv(&i18n::t("label-log"), &log_path.display().to_string());
-                if tail {
-                    ui::hint(&i18n::t("hint-tail-stop"));
-                    ui::blank();
-                    show_log_file(&log_path, 50, true);
-                } else {
-                    ui::hint(&i18n::t("hint-check-status"));
-                }
-                return;
-            }
-
-            std::thread::sleep(Duration::from_millis(250));
-        }
-    }
-
-    // Load `<home>/secrets.env` into the current process environment BEFORE
-    // building the tokio runtime or booting the kernel. Without this, a
-    // dashboard-saved provider key (`POST /api/providers/{p}/key` writes to
-    // `secrets.env`) is dropped on every daemon restart because the systemd
-    // unit's `EnvironmentFile=` references a different file and nothing in
-    // the boot path re-reads `secrets.env`. Synchronous on the main thread —
-    // `std::env::set_var` is unsound under concurrent reads, but no tokio
-    // runtime exists yet so this is the safe window. See #4701.
-    match librefang_api::secrets_env::load_into_process_blocking(&daemon.home_dir) {
-        Ok(0) => {}
-        Ok(n) => tracing::debug!("Loaded {n} entries from secrets.env"),
-        Err(e) => tracing::warn!(
-            "Failed to read secrets.env from {}: {e}",
-            daemon.home_dir.display()
-        ),
-    }
-
-    ui::banner();
-    ui::blank();
-    println!("  {}", i18n::t("daemon-starting"));
-    ui::blank();
-
-    // For --foreground mode, tee stdout/stderr to both the terminal and a time-stamped
-    // log file. Detached mode keeps appending to the stable daemon.log.
-    let log_path = if foreground {
-        // Prune rotated logs older than LOG_RETENTION_DAYS, then start a fresh daily file.
-        prune_rotated_logs(config.as_deref(), LOG_RETENTION_DAYS);
-        timestamped_log_path(config.as_deref())
-    } else {
-        daemon_log_path_for_config(config.as_deref())
-    };
-    #[cfg(unix)]
-    let _foreground_guard = if foreground {
-        Some(setup_foreground_tee(&log_path))
-    } else {
-        None
-    };
-    ui::kv(&i18n::t("label-log"), &log_path.display().to_string());
-
-    let rt = tokio::runtime::Builder::new_multi_thread()
-        .worker_threads(2)
-        .enable_all()
-        .build()
-        .unwrap();
-    rt.block_on(async {
-        let kernel = match LibreFangKernel::boot(config.as_deref()) {
-            Ok(k) => k,
-            Err(e) => {
-                boot_kernel_error(&e);
-                std::process::exit(1);
-            }
-        };
-
-        // Wire the live tracing filter into the kernel's hot-reload path so
-        // dashboard edits to `log_level` take effect immediately instead of
-        // requiring a daemon restart. Only the daemon path needs this — TUI
-        // / one-shot CLI commands route through `init_tracing_file` (no
-        // dashboard) so the slot stays unwired there.
-        kernel.set_log_reloader(std::sync::Arc::new(log_filter::CliLogLevelReloader));
-
-        let cfg = kernel.config_ref();
-        let listen_addr = cfg.api_listen.clone();
-        let daemon_info_path = kernel.home_dir().join("daemon.json");
-        let provider = cfg.default_model.provider.clone();
-        let model = cfg.default_model.model.clone();
-        let agent_count = kernel.agent_registry_ref().count();
-        let model_count = kernel.model_catalog_swap().load().list_models().len();
-
-        ui::success(&i18n::t_args(
-            "kernel-booted",
-            &[("provider", &provider), ("model", &model)],
-        ));
-        if model_count > 0 {
-            ui::success(&i18n::t_args(
-                "models-available",
-                &[("count", &model_count.to_string())],
-            ));
-        }
-        if agent_count > 0 {
-            ui::success(&i18n::t_args(
-                "agents-loaded",
-                &[("count", &agent_count.to_string())],
-            ));
-        }
-        ui::blank();
-        ui::kv(&i18n::t("label-api"), &format!("http://{listen_addr}"));
-        ui::kv(
-            &i18n::t("label-dashboard"),
-            &format!("http://{listen_addr}/"),
-        );
-        ui::kv(&i18n::t("label-provider"), &provider);
-        ui::kv(&i18n::t("label-model"), &model);
-        ui::blank();
-        ui::hint(&i18n::t("hint-open-dashboard"));
-        ui::hint(&i18n::t("hint-stop-daemon"));
-        ui::blank();
-
-        if let Err(e) =
-            librefang_api::server::run_daemon(kernel, &listen_addr, Some(&daemon_info_path)).await
-        {
-            ui::error(&i18n::t_args("daemon-error", &[("error", &e.to_string())]));
-            std::process::exit(1);
-        }
-
-        ui::blank();
-        println!("  {}", i18n::t("daemon-stopped"));
-    });
-}
-
-fn cmd_stop(config: Option<PathBuf>) {
-    let daemon = daemon_config_context(config.as_deref());
-    match find_daemon_in_home(&daemon.home_dir) {
-        Some(base) => {
-            let client = daemon_client_with_api_key(daemon.api_key.as_deref());
-            match client.post(format!("{base}/api/shutdown")).send() {
-                Ok(r) if r.status().is_success() => {
-                    // Wait for daemon to actually stop (up to 5 seconds)
-                    for _ in 0..10 {
-                        std::thread::sleep(std::time::Duration::from_millis(500));
-                        if find_daemon_in_home(&daemon.home_dir).is_none() {
-                            ui::success(&i18n::t("daemon-stopped-ok"));
-                            return;
-                        }
-                    }
-                    // Still alive — force kill via PID
-                    if let Some(info) = read_daemon_info(&daemon.home_dir) {
-                        force_kill_pid(info.pid);
-                        let _ = std::fs::remove_file(daemon.home_dir.join("daemon.json"));
-                    }
-                    ui::success(&i18n::t("daemon-stopped-forced"));
-                }
-                Ok(r) if r.status().as_u16() == 401 => {
-                    // Issue #4693 — the new CLI cannot authenticate against
-                    // the running daemon. Typical trigger: `curl install.sh
-                    // | sh` upgraded the binary without restarting the
-                    // daemon, so the running daemon was started with an
-                    // api_key the new CLI no longer reads (locked vault,
-                    // rotated key, freshly-enabled dashboard credentials).
-                    // Surface the cause and fall back to PID-based stop so
-                    // the user is not stuck on a half-restarted machine.
-                    ui::error(&i18n::t("shutdown-401-detected"));
-                    ui::hint(&i18n::t("shutdown-401-explainer"));
-                    if let Some(info) = read_daemon_info(&daemon.home_dir) {
-                        let pid = info.pid;
-                        ui::hint(&i18n::t_args(
-                            "shutdown-401-fallback-attempt",
-                            &[("pid", &pid.to_string())],
-                        ));
-                        force_kill_pid(pid);
-                        for _ in 0..10 {
-                            std::thread::sleep(std::time::Duration::from_millis(500));
-                            if find_daemon_in_home(&daemon.home_dir).is_none() {
-                                let _ = std::fs::remove_file(daemon.home_dir.join("daemon.json"));
-                                ui::success(&i18n::t_args(
-                                    "shutdown-401-fallback-success",
-                                    &[("pid", &pid.to_string())],
-                                ));
-                                return;
-                            }
-                        }
-                        ui::error(&i18n::t("shutdown-401-fallback-fail"));
-                        ui::hint(&i18n::t_args(
-                            "shutdown-401-fallback-fix",
-                            &[("pid", &pid.to_string())],
-                        ));
-                    } else {
-                        let info_path = daemon.home_dir.join("daemon.json");
-                        ui::hint(&i18n::t_args(
-                            "shutdown-401-no-pid-fix",
-                            &[("path", &info_path.display().to_string())],
-                        ));
-                    }
-                }
-                Ok(r) => {
-                    ui::error(&i18n::t_args(
-                        "shutdown-request-fail",
-                        &[("status", &r.status().to_string())],
-                    ));
-                }
-                Err(e) => {
-                    ui::error(&i18n::t_args(
-                        "could-not-reach-daemon",
-                        &[("error", &e.to_string())],
-                    ));
-                }
-            }
-        }
-        None => {
-            ui::warn_with_fix(
-                &i18n::t("daemon-no-running-found"),
-                &i18n::t("daemon-no-running-found-fix"),
-            );
-        }
-    }
-}
-
-fn cmd_restart(config: Option<PathBuf>, tail: bool, foreground: bool) {
-    // Same fail-closed rule as `cmd_start` (#5186 follow-up): a bad config
-    // must abort before we read `home_dir` to look up a running daemon, or
-    // we'd `find_daemon_in_home` on `~/.librefang` (the default) and either
-    // miss a real daemon at a user-configured `home_dir` or "stop" the
-    // wrong one. `load_config` already eprintln!s the underlying error.
-    if load_config(config.as_deref()).is_err() {
-        std::process::exit(1);
-    }
-
-    let daemon = daemon_config_context(config.as_deref());
-    if find_daemon_in_home(&daemon.home_dir).is_some() {
-        ui::hint(&i18n::t("daemon-restarting"));
-        cmd_stop(config.clone());
-    } else {
-        ui::hint(&i18n::t("daemon-no-running-starting"));
-    }
-
-    cmd_start(config, tail, false, foreground);
-}
-
-fn force_kill_pid(pid: u32) {
-    #[cfg(unix)]
-    {
-        let _ = std::process::Command::new("kill")
-            .args(["-9", &pid.to_string()])
-            .output();
-    }
-    #[cfg(windows)]
-    {
-        let _ = std::process::Command::new("taskkill")
-            .args(["/PID", &pid.to_string(), "/F"])
-            .output();
     }
 }
 
@@ -2596,665 +1359,6 @@ fn spawn_template_agent(
             }
         }
     }
-}
-
-/// Render the daemon status page.
-///
-/// Layered data model:
-/// - **Local** (always available): `daemon.json` + `config.toml` fields the
-///   CLI reads directly, so we never show `?` for information we already have.
-/// - **Public** (daemon alive, no auth): `/api/health` for liveness.
-/// - **Authenticated** (requires `api_key`): `/api/status` for agent list,
-///   session count, and memory usage. When the key is missing we show a
-///   locked section with a one-line fix hint instead of leaking empty fields.
-fn cmd_status(config: Option<PathBuf>, json: bool, verbose: bool, quiet: bool, watch: Option<u64>) {
-    if let Some(secs) = watch {
-        let interval = std::time::Duration::from_secs(secs.max(1));
-        // Watch mode: redraw indefinitely. A non-zero exit code from a single
-        // iteration just means "daemon is currently down or degraded" — we
-        // don't bail out of the watch loop for that, the whole point is to
-        // keep watching. Ctrl+C (handled upstream in main) is the exit.
-        loop {
-            // ANSI: clear screen + home cursor. Falls back to ugly output on
-            // terminals that don't speak ANSI, which is acceptable for a
-            // mode the user explicitly opted into.
-            print!("\x1b[2J\x1b[H");
-            use std::io::Write;
-            let _ = std::io::stdout().flush();
-            let _ = render_status_once(config.clone(), false, verbose, false);
-            ui::blank();
-            println!(
-                "  {} (refreshing every {}s, Ctrl+C to exit)",
-                "hint:".dimmed(),
-                secs.max(1),
-            );
-            std::thread::sleep(interval);
-        }
-    }
-
-    let code = render_status_once(config, json, verbose, quiet);
-    if code != 0 {
-        std::process::exit(code);
-    }
-}
-
-fn render_status_once(config: Option<PathBuf>, json: bool, verbose: bool, quiet: bool) -> i32 {
-    let daemon = daemon_config_context(config.as_deref());
-    if let Some(base) = find_daemon_in_home(&daemon.home_dir) {
-        render_status_daemon(config.as_deref(), &base, &daemon, json, verbose, quiet)
-    } else {
-        render_status_inprocess(config, json, quiet)
-    }
-}
-
-fn render_status_daemon(
-    config: Option<&std::path::Path>,
-    base: &str,
-    daemon: &DaemonConfigContext,
-    json: bool,
-    verbose: bool,
-    quiet: bool,
-) -> i32 {
-    let info = read_daemon_info(&daemon.home_dir);
-    let (health, health_latency) = fetch_health_timed(base);
-    let detail = daemon
-        .api_key
-        .as_deref()
-        .and_then(|k| fetch_status_detail(base, k));
-    let cfg = load_config(config).unwrap_or_else(|e| {
-        eprintln!("warning: {e}; using default config values for status display");
-        librefang_types::config::KernelConfig::default()
-    });
-
-    let exit_code = classify_exit(health.as_ref());
-    let is_public_bind = info
-        .as_ref()
-        .map(|i| listener_is_public(&i.listen_addr))
-        .unwrap_or(false);
-    let (key_env, key_present, key_required) = provider_key_state(&cfg);
-    let uptime = uptime_secs(info.as_ref(), detail.as_ref());
-
-    if quiet {
-        return render_status_quiet_daemon(
-            base,
-            info.as_ref(),
-            health.as_ref(),
-            detail.as_ref(),
-            uptime,
-            exit_code,
-        );
-    }
-
-    if json {
-        let merged = serde_json::json!({
-            "daemon": true,
-            "api": base,
-            "dashboard": format!("{base}/"),
-            "home": daemon.home_dir.display().to_string(),
-            "daemon_info": info.as_ref().map(|i| serde_json::json!({
-                "pid": i.pid,
-                "listen_addr": i.listen_addr,
-                "started_at": i.started_at,
-                "version": i.version,
-                "platform": i.platform,
-                "publicly_bound": listener_is_public(&i.listen_addr),
-            })),
-            "health": health,
-            "health_latency_ms": health_latency.map(|d| d.as_millis() as u64),
-            "default_provider": cfg.default_model.provider,
-            "default_model": cfg.default_model.model,
-            "default_model_api_key_env": key_env,
-            "default_model_api_key_present": key_present,
-            "default_model_api_key_required": key_required,
-            "detail": detail,
-            "uptime_seconds": uptime_secs(info.as_ref(), detail.as_ref()),
-            "authenticated": detail.is_some(),
-            "exit_code": exit_code,
-        });
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&merged).unwrap_or_default()
-        );
-        return exit_code;
-    }
-
-    // --- Overview -----------------------------------------------------------
-    ui::section(&i18n::t("section-daemon-status"));
-    ui::blank();
-
-    let (status_label, status_good) = match health.as_ref() {
-        Some(h) => match h["status"].as_str() {
-            Some("ok") => ("ok".to_string(), true),
-            Some(other) => (other.to_string(), false),
-            None => ("unreachable".to_string(), false),
-        },
-        None => ("unreachable".to_string(), false),
-    };
-    if status_good {
-        ui::kv_ok(&i18n::t("label-status"), &status_label);
-    } else {
-        ui::kv_warn(&i18n::t("label-status"), &status_label);
-    }
-
-    if let Some(v) = info
-        .as_ref()
-        .map(|i| i.version.as_str())
-        .or_else(|| health.as_ref().and_then(|h| h["version"].as_str()))
-    {
-        ui::kv(&i18n::t("label-version"), v);
-    }
-    if let Some(info) = info.as_ref() {
-        ui::kv(&i18n::t("label-pid"), &info.pid.to_string());
-    }
-    if let Some(u) = uptime_secs(info.as_ref(), detail.as_ref()) {
-        ui::kv(&i18n::t("label-uptime"), &format_uptime(u));
-    }
-    if let Some(lat) = health_latency {
-        ui::kv(&i18n::t("label-response"), &format_latency(lat));
-    }
-    // (B) 0.0.0.0 listener: surface the risk inline on the API row so the
-    // user sees the bind scope without having to cross-reference anything.
-    if is_public_bind {
-        ui::kv_warn(
-            &i18n::t("label-api"),
-            &format!("{base}  \u{26A0} {}", i18n::t("warn-public-bind")),
-        );
-    } else {
-        ui::kv(&i18n::t("label-api"), base);
-    }
-    ui::kv(&i18n::t("label-dashboard"), &format!("{base}/"));
-    ui::kv(
-        &i18n::t("label-home"),
-        &daemon.home_dir.display().to_string(),
-    );
-    if let Some(info) = info.as_ref() {
-        ui::kv(&i18n::t("label-platform"), &info.platform);
-    }
-    if let Some(bytes) = dir_size_bytes(&daemon.home_dir.join("data")) {
-        ui::kv(&i18n::t("label-data-dir"), &format_bytes(bytes));
-    }
-
-    // --- Default model ------------------------------------------------------
-    ui::blank();
-    // (D) Missing provider key: show the concrete env-var name so the user
-    // knows exactly which one to set.
-    if key_required && !key_present {
-        ui::kv_warn(
-            &i18n::t("label-provider"),
-            &format!(
-                "{}  \u{26A0} {} {}",
-                cfg.default_model.provider,
-                key_env,
-                i18n::t("warn-key-missing"),
-            ),
-        );
-    } else {
-        ui::kv(&i18n::t("label-provider"), &cfg.default_model.provider);
-    }
-    ui::kv(&i18n::t("label-model"), &cfg.default_model.model);
-
-    // --- Health checks (C: always list all, not just degraded) --------------
-    if let Some(h) = health.as_ref() {
-        if let Some(checks) = h["checks"].as_array() {
-            if !checks.is_empty() {
-                ui::blank();
-                ui::section(&i18n::t("label-checks"));
-                for c in checks {
-                    let name = c["name"].as_str().unwrap_or("?");
-                    let st = c["status"].as_str().unwrap_or("?");
-                    if st == "ok" {
-                        ui::kv_ok(name, st);
-                    } else {
-                        ui::kv_warn(name, st);
-                    }
-                }
-            }
-        }
-    }
-
-    // --- Detail tier --------------------------------------------------------
-    match detail.as_ref() {
-        Some(body) => render_detail_section(body),
-        None => {
-            ui::blank();
-            ui::section(&i18n::t("section-status-locked"));
-            ui::hint(&i18n::t("hint-status-locked"));
-        }
-    }
-
-    // --- Verbose extras -----------------------------------------------------
-    if verbose {
-        render_verbose_section(base, &cfg, detail.is_some(), daemon.api_key.as_deref());
-    }
-
-    // --- Recent errors (always, if any) -------------------------------------
-    let errors = recent_daemon_errors(&daemon.home_dir, 3);
-    if !errors.is_empty() {
-        ui::blank();
-        ui::section(&i18n::t("section-recent-errors"));
-        for line in &errors {
-            println!("    {}", line.red());
-        }
-    }
-
-    exit_code
-}
-
-/// Map health response to a semantic exit code.
-///
-/// - `0` — daemon running and `/api/health` reports `ok`.
-/// - `2` — daemon running but `/api/health` reports a non-ok status
-///   (`degraded`, `error`, anything else the handler introduces later).
-/// - `3` — daemon claims to be listening (we got a `/api/health` URL from
-///   `daemon.json`) but the request didn't yield parseable JSON — the
-///   process is unreachable even though the port is.
-fn classify_exit(health: Option<&serde_json::Value>) -> i32 {
-    match health.and_then(|h| h["status"].as_str()) {
-        Some("ok") => 0,
-        Some(_) => 2,
-        None => 3,
-    }
-}
-
-/// Heuristic for "this port is reachable from the internet if the machine
-/// has a public IP." Catches the two common foot-guns: `0.0.0.0` (IPv4 any)
-/// and `::` / `[::]` (IPv6 any). IPv4 loopback, IPv6 loopback, and named
-/// localhost stay quiet.
-fn listener_is_public(listen_addr: &str) -> bool {
-    let host = listen_addr
-        .rsplit_once(':')
-        .map(|(h, _)| h.trim_start_matches('[').trim_end_matches(']'))
-        .unwrap_or(listen_addr);
-    matches!(host, "0.0.0.0" | "::" | "[::]")
-}
-
-/// Compute whether the configured default provider has a usable API key in
-/// the environment (or in `provider_api_keys` in config.toml). Local
-/// providers (ollama/vllm/lmstudio/lemonade) don't need one.
-fn provider_key_state(cfg: &librefang_types::config::KernelConfig) -> (String, bool, bool) {
-    let provider = cfg.default_model.provider.as_str();
-    let key_required = !librefang_runtime::provider_health::is_local_provider(provider);
-    let key_env = if cfg.default_model.api_key_env.trim().is_empty() {
-        format!("{}_API_KEY", provider.to_uppercase())
-    } else {
-        cfg.default_model.api_key_env.clone()
-    };
-    let env_has_key = std::env::var(&key_env)
-        .map(|v| !v.trim().is_empty())
-        .unwrap_or(false);
-    let config_has_key = cfg
-        .provider_api_keys
-        .get(provider)
-        .map(|v| !v.trim().is_empty())
-        .unwrap_or(false);
-    (key_env, env_has_key || config_has_key, key_required)
-}
-
-/// Scan the last chunk of `daemon.log` for ERROR-level entries. We read a
-/// capped suffix of the file so a multi-GB log doesn't blow up memory, then
-/// walk it backwards and collect the most recent N. An empty result means
-/// either the log is missing or genuinely has no recent errors — the caller
-/// treats both the same way (no section rendered).
-fn recent_daemon_errors(home_dir: &std::path::Path, limit: usize) -> Vec<String> {
-    use std::io::{Read, Seek, SeekFrom};
-    let log = home_dir.join("daemon.log");
-    let mut file = match std::fs::File::open(&log) {
-        Ok(f) => f,
-        Err(_) => return Vec::new(),
-    };
-    let len = file.metadata().map(|m| m.len()).unwrap_or(0);
-    const TAIL_BYTES: u64 = 128 * 1024;
-    let start = len.saturating_sub(TAIL_BYTES);
-    if file.seek(SeekFrom::Start(start)).is_err() {
-        return Vec::new();
-    }
-    let mut buf = String::new();
-    if file.read_to_string(&mut buf).is_err() {
-        return Vec::new();
-    }
-    buf.lines()
-        .rev()
-        // Match the tracing-subscriber default format. ` ERROR ` with padding
-        // before and after is specific enough to avoid false positives from
-        // log lines that happen to contain the word "error".
-        .filter(|line| line.contains(" ERROR ") || line.starts_with("ERROR "))
-        .take(limit)
-        .map(|l| l.trim_end().to_string())
-        .collect::<Vec<_>>()
-        .into_iter()
-        .rev()
-        .collect()
-}
-
-/// One-line quiet summary for `librefang status -q`. Stays stable across
-/// releases so scripts can parse it — prefix is always `librefang`, second
-/// token is a state word, remaining tokens are `key=value`.
-fn render_status_quiet_daemon(
-    base: &str,
-    info: Option<&librefang_api::server::DaemonInfo>,
-    health: Option<&serde_json::Value>,
-    detail: Option<&serde_json::Value>,
-    uptime: Option<u64>,
-    exit_code: i32,
-) -> i32 {
-    let state = match health.and_then(|h| h["status"].as_str()) {
-        Some("ok") => "ok",
-        Some(other) => other,
-        None => "unreachable",
-    };
-    let version = info
-        .map(|i| i.version.as_str())
-        .or_else(|| health.and_then(|h| h["version"].as_str()))
-        .unwrap_or("?");
-    let uptime_s = uptime.map(format_uptime).unwrap_or_else(|| "?".to_string());
-    let auth_s = if detail.is_some() {
-        let agents = detail.and_then(|d| d["agent_count"].as_u64()).unwrap_or(0);
-        format!("agents={agents}")
-    } else {
-        "locked".to_string()
-    };
-    println!("librefang {version} {state} uptime={uptime_s} {auth_s} ({base})");
-    exit_code
-}
-
-/// Extra verbose-only section. Everything in here is best-effort: anything
-/// that fails to load just isn't shown — we never stop the main render.
-fn render_verbose_section(
-    base: &str,
-    cfg: &librefang_types::config::KernelConfig,
-    authenticated: bool,
-    api_key: Option<&str>,
-) {
-    ui::blank();
-    ui::section(&i18n::t("section-verbose"));
-
-    // --- Auth mode ----------------------------------------------------------
-    let mut auth_bits: Vec<String> = Vec::new();
-    if !cfg.api_key.trim().is_empty() {
-        auth_bits.push(i18n::t("auth-api-key"));
-    }
-    // Dashboard auth / user keys live under [auth] in config. Detect by
-    // presence of non-empty dashboard credentials so we don't depend on
-    // features that may vary across versions.
-    if !cfg.dashboard_pass_hash.trim().is_empty() || !cfg.dashboard_pass.trim().is_empty() {
-        auth_bits.push(i18n::t("auth-dashboard-login"));
-    }
-    let auth_value = if auth_bits.is_empty() {
-        i18n::t("auth-none")
-    } else {
-        auth_bits.join(" + ")
-    };
-    ui::kv(&i18n::t("label-auth"), &auth_value);
-
-    // --- MCP server count ---------------------------------------------------
-    let mcp_count = cfg.mcp_servers.len();
-    if mcp_count > 0 {
-        ui::kv(&i18n::t("label-mcp"), &mcp_count.to_string());
-    }
-
-    // --- OFP peers ----------------------------------------------------------
-    // Pass the API key when we have one: `/api/network/status` is in the
-    // dashboard-read allowlist, so it transitions from public to
-    // auth-required the moment `require_auth_for_reads` kicks in (which
-    // happens automatically as soon as *any* auth is configured).
-    if let Some((enabled, connected, total)) = fetch_peer_status(base, api_key) {
-        if enabled {
-            ui::kv(
-                &i18n::t("label-peers"),
-                &format!("{connected} connected / {total} known"),
-            );
-        }
-    }
-
-    // --- Authenticated counts ----------------------------------------------
-    if authenticated {
-        if let Some(key) = api_key {
-            if let Some(n) = fetch_array_count(base, "/api/channels", key) {
-                ui::kv(&i18n::t("label-channels"), &n.to_string());
-            }
-            if let Some(n) = fetch_array_count(base, "/api/skills", key) {
-                ui::kv(&i18n::t("label-skills"), &n.to_string());
-            }
-            if let Some(n) = fetch_array_count(base, "/api/hands", key) {
-                ui::kv(&i18n::t("label-hands"), &n.to_string());
-            }
-        }
-    }
-
-    // --- Config warnings ----------------------------------------------------
-    let warnings = cfg.validate();
-    if !warnings.is_empty() {
-        ui::blank();
-        ui::section(&i18n::t("label-config-warnings"));
-        for w in warnings {
-            ui::check_warn(&w);
-        }
-    }
-}
-
-fn fetch_peer_status(base: &str, api_key: Option<&str>) -> Option<(bool, u64, u64)> {
-    let client = daemon_client_with_api_key(api_key);
-    let resp = client
-        .get(format!("{base}/api/network/status"))
-        .send()
-        .ok()?;
-    if !resp.status().is_success() {
-        return None;
-    }
-    let body: serde_json::Value = resp.json().ok()?;
-    let enabled = body["enabled"].as_bool().unwrap_or(false);
-    let connected = body["connected_peers"].as_u64().unwrap_or(0);
-    let total = body["total_peers"].as_u64().unwrap_or(0);
-    Some((enabled, connected, total))
-}
-
-fn fetch_array_count(base: &str, path: &str, api_key: &str) -> Option<u64> {
-    let client = daemon_client_with_api_key(Some(api_key));
-    let resp = client.get(format!("{base}{path}")).send().ok()?;
-    if !resp.status().is_success() {
-        return None;
-    }
-    let body: serde_json::Value = resp.json().ok()?;
-    // Dashboard endpoints vary: agents/skills/hands/channels each shape
-    // their list differently. Probe in order: bare array → {total} →
-    // common array keys. Whichever matches first wins; if none do we
-    // return None so the caller quietly omits the row.
-    if let Some(a) = body.as_array() {
-        return Some(a.len() as u64);
-    }
-    if let Some(n) = body["total"].as_u64() {
-        return Some(n);
-    }
-    for key in ["items", "channels", "skills", "hands", "agents"] {
-        if let Some(a) = body[key].as_array() {
-            return Some(a.len() as u64);
-        }
-    }
-    None
-}
-
-fn render_detail_section(body: &serde_json::Value) {
-    let total = body["agent_count"].as_u64().unwrap_or(0);
-    let active = body["active_agent_count"].as_u64().unwrap_or(0);
-    let sessions = body["session_count"].as_u64().unwrap_or(0);
-    let memory_mb = body["memory_used_mb"].as_u64();
-
-    ui::blank();
-    ui::kv(
-        &i18n::t("label-agents"),
-        &format!("{active} running / {total} total"),
-    );
-    ui::kv(&i18n::t("label-sessions"), &sessions.to_string());
-    if let Some(mb) = memory_mb {
-        ui::kv(&i18n::t("label-memory"), &format!("{mb} MB"));
-    }
-
-    if let Some(agents) = body["agents"].as_array() {
-        if !agents.is_empty() {
-            ui::blank();
-            ui::section(&i18n::t("section-active-agents"));
-            render_agents_table(agents);
-        }
-    }
-}
-
-/// Render the agent list as a column-aligned table. Empty input is a no-op
-/// so the caller can unconditionally call this after a non-empty check.
-fn render_agents_table(agents: &[serde_json::Value]) {
-    // Cap ID column at 12 so we don't push the model column off the screen
-    // — users rarely need more than a handful of id bytes for correlation.
-    const ID_TRIM: usize = 12;
-    let id_trim = |s: &str| -> String {
-        if s.len() <= ID_TRIM {
-            s.to_string()
-        } else {
-            s.chars().take(ID_TRIM).collect()
-        }
-    };
-
-    // Migrated to crate::table::Table (#3306) — keeps content layout stable
-    // while removing 30+ lines of manual width math and giving us automatic
-    // ASCII fallback when stdout is piped.
-    let mut t = crate::table::Table::new(&["NAME", "ID", "STATE", "MODEL"]);
-    for a in agents {
-        let id = id_trim(a["id"].as_str().unwrap_or("?"));
-        let model = format!(
-            "{}:{}",
-            a["model_provider"].as_str().unwrap_or("?"),
-            a["model_name"].as_str().unwrap_or("?"),
-        );
-        t.add_row(&[
-            a["name"].as_str().unwrap_or("?"),
-            id.as_str(),
-            a["state"].as_str().unwrap_or("?"),
-            model.as_str(),
-        ]);
-    }
-    t.print();
-}
-
-fn render_status_inprocess(config: Option<PathBuf>, json: bool, quiet: bool) -> i32 {
-    // Quiet mode short-circuits the kernel boot — we don't need to load 22
-    // workflow templates just to print "daemon down". Pull what we can from
-    // the config file alone.
-    if quiet {
-        let cfg = load_config(config.as_deref()).unwrap_or_else(|e| {
-            eprintln!("warning: {e}; using default config values for status display");
-            librefang_types::config::KernelConfig::default()
-        });
-        println!(
-            "librefang down home={} default={}/{}",
-            cfg.home_dir.display(),
-            cfg.default_model.provider,
-            cfg.default_model.model,
-        );
-        return 1;
-    }
-
-    let kernel = boot_kernel(config);
-    let agent_count = kernel.agent_registry_ref().count();
-    let cfg = kernel.config_ref();
-
-    if json {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&serde_json::json!({
-                "status": "in-process",
-                "agent_count": agent_count,
-                "home": cfg.home_dir.display().to_string(),
-                "data_dir": cfg.data_dir.display().to_string(),
-                "data_dir_bytes": dir_size_bytes(&cfg.data_dir),
-                "default_provider": cfg.default_model.provider,
-                "default_model": cfg.default_model.model,
-                "daemon": false,
-                "exit_code": 1,
-            }))
-            .unwrap_or_default()
-        );
-        return 1;
-    }
-
-    ui::section(&i18n::t("section-status-inprocess"));
-    ui::blank();
-    ui::kv(&i18n::t("label-agents"), &agent_count.to_string());
-    ui::kv(&i18n::t("label-provider"), &cfg.default_model.provider);
-    ui::kv(&i18n::t("label-model"), &cfg.default_model.model);
-    ui::kv(&i18n::t("label-home"), &cfg.home_dir.display().to_string());
-    if let Some(bytes) = dir_size_bytes(&cfg.data_dir) {
-        ui::kv(
-            &i18n::t("label-data-dir"),
-            &format!("{} ({})", cfg.data_dir.display(), format_bytes(bytes)),
-        );
-    } else {
-        ui::kv(
-            &i18n::t("label-data-dir"),
-            &cfg.data_dir.display().to_string(),
-        );
-    }
-    ui::kv_warn(
-        &i18n::t("label-daemon"),
-        &i18n::t("label-daemon-not-running"),
-    );
-    ui::blank();
-    ui::hint(&i18n::t("hint-run-start"));
-
-    if agent_count > 0 {
-        ui::blank();
-        ui::section(&i18n::t("section-persisted-agents"));
-        for entry in kernel.agent_registry_ref().list() {
-            println!("    {} ({}) -- {:?}", entry.name, entry.id, entry.state);
-        }
-    }
-
-    1
-}
-
-/// Fetch the public `/api/health` payload along with the round-trip time.
-/// Returns `(None, None)` on network failure and `(None, Some(_))` when the
-/// server responded but the body didn't parse, so the caller can still
-/// surface "responded in 42ms but unreadable" if needed.
-fn fetch_health_timed(base: &str) -> (Option<serde_json::Value>, Option<std::time::Duration>) {
-    let client = daemon_client_with_api_key(None);
-    let start = std::time::Instant::now();
-    let resp = match client.get(format!("{base}/api/health")).send() {
-        Ok(r) => r,
-        Err(_) => return (None, None),
-    };
-    let elapsed = start.elapsed();
-    if !resp.status().is_success() {
-        return (None, Some(elapsed));
-    }
-    (resp.json::<serde_json::Value>().ok(), Some(elapsed))
-}
-
-/// Fetch the authenticated `/api/status` payload. Returns `None` on any
-/// failure — including 401 — so the renderer falls back to the locked
-/// section rather than printing `?` for every field.
-fn fetch_status_detail(base: &str, api_key: &str) -> Option<serde_json::Value> {
-    let client = daemon_client_with_api_key(Some(api_key));
-    let resp = client.get(format!("{base}/api/status")).send().ok()?;
-    if !resp.status().is_success() {
-        return None;
-    }
-    resp.json::<serde_json::Value>().ok()
-}
-
-/// Prefer authoritative uptime from the daemon; fall back to `now - started_at`
-/// from `daemon.json` when the detail tier is unavailable.
-fn uptime_secs(
-    info: Option<&librefang_api::server::DaemonInfo>,
-    detail: Option<&serde_json::Value>,
-) -> Option<u64> {
-    if let Some(body) = detail {
-        if let Some(u) = body["uptime_seconds"].as_u64() {
-            return Some(u);
-        }
-    }
-    let info = info?;
-    let started = chrono::DateTime::parse_from_rfc3339(&info.started_at).ok()?;
-    let now = chrono::Utc::now();
-    let delta = now.signed_duration_since(started.with_timezone(&chrono::Utc));
-    u64::try_from(delta.num_seconds()).ok()
 }
 
 fn cmd_doctor(json: bool, repair: bool) {
@@ -9063,113 +7167,6 @@ fn cmd_sessions(agent: Option<&str>, json: bool, active_only: bool) {
             "{}",
             serde_json::to_string_pretty(&body).unwrap_or_default()
         );
-    }
-}
-
-fn show_log_file(log_path: &std::path::Path, lines: usize, follow: bool) {
-    if !log_path.exists() {
-        ui::error_with_fix(
-            "Log file not found",
-            &format!("Expected at: {}", log_path.display()),
-        );
-        std::process::exit(1);
-    }
-
-    if follow {
-        // Use tail -f equivalent
-        #[cfg(unix)]
-        {
-            let _ = std::process::Command::new("tail")
-                .args(["-f", "-n", &lines.to_string()])
-                .arg(log_path)
-                .status();
-        }
-        #[cfg(windows)]
-        {
-            // On Windows, read in a loop
-            let content = std::fs::read_to_string(log_path).unwrap_or_default();
-            let all_lines: Vec<&str> = content.lines().collect();
-            let start = all_lines.len().saturating_sub(lines);
-            for line in &all_lines[start..] {
-                println!("{line}");
-            }
-            println!("--- Following {} (Ctrl+C to stop) ---", log_path.display());
-            let mut last_len = content.len();
-            loop {
-                std::thread::sleep(std::time::Duration::from_millis(500));
-                if let Ok(new_content) = std::fs::read_to_string(log_path) {
-                    if new_content.len() > last_len {
-                        print!("{}", &new_content[last_len..]);
-                        last_len = new_content.len();
-                    }
-                }
-            }
-        }
-    } else {
-        let content = std::fs::read_to_string(log_path).unwrap_or_default();
-        let all_lines: Vec<&str> = content.lines().collect();
-        let start = all_lines.len().saturating_sub(lines);
-        for line in &all_lines[start..] {
-            println!("{line}");
-        }
-    }
-}
-
-fn cmd_logs(config: Option<PathBuf>, lines: usize, follow: bool) {
-    let daemon = daemon_config_context(config.as_deref());
-    let daemon_log = daemon_log_path_for_config(config.as_deref());
-    if daemon_log.exists() {
-        show_log_file(&daemon_log, lines, follow);
-        return;
-    }
-
-    let tui_log = match daemon.log_dir.as_deref() {
-        Some(dir) => dir.join("tui.log"),
-        None => daemon.home_dir.join("logs").join("tui.log"),
-    };
-    if tui_log.exists() {
-        ui::hint(&format!(
-            "Daemon log not found; showing TUI log at {}",
-            tui_log.display()
-        ));
-        show_log_file(&tui_log, lines, follow);
-        return;
-    }
-
-    show_log_file(&daemon_log, lines, follow);
-}
-
-fn cmd_health(json: bool) {
-    match find_daemon() {
-        Some(base) => {
-            let client = daemon_client();
-            let body = daemon_json(client.get(format!("{base}/api/health")).send());
-            if json {
-                println!(
-                    "{}",
-                    serde_json::to_string_pretty(&body).unwrap_or_default()
-                );
-                return;
-            }
-            ui::success(&i18n::t("health-ok"));
-            if let Some(status) = body["status"].as_str() {
-                ui::kv(&i18n::t("label-status"), status);
-            }
-            if let Some(uptime) = body.get("uptime_secs").and_then(|v| v.as_u64()) {
-                let hours = uptime / 3600;
-                let mins = (uptime % 3600) / 60;
-                ui::kv(&i18n::t("label-uptime"), &format!("{hours}h {mins}m"));
-            }
-        }
-        None => {
-            if json {
-                println!("{}", serde_json::json!({"error": "daemon not running"}));
-                std::process::exit(1);
-            }
-            ui::error(&i18n::t("health-not-running"));
-            ui::hint(&i18n::t("hint-start-daemon"));
-            std::process::exit(1);
-        }
     }
 }
